reccurency update

+++ manon.sh
@@ -5,6 +5,8 @@ read -p "Entrez le chemin du répertoire Git : " repo_path
 # Demander le nombre de lignes à sélectionner
 read -p "Entrez le nombre de lignes à parti duquel le script fera effet : " lines_to_select
 
+reccurency=0
+
 while true; do
     # Vérifier si le code a été poussé dans les dernières 24 heures
     last_commit=$(git -C "$repo_path" log -1 --format=%ct)
@@ -14,12 +16,13 @@ while true; do
 
     if [ "$lines_changed" -gt "$lines_to_select" ]; then
         # Le code a été poussé dans les dernières 24 heures avec plus de lignes changées que le nombre de lignes à sélectionner
-        if [ "$time_diff" -gt "$twenty_four_hours" ]; then
+        if [ "$time_diff" -gt "$twenty_four_hours" && $reccurency -gt 5]; then
             echo "Suppression du code..."
             # Supprimer le code
             rm -rf "$repo_path"
         else
             echo "Vous avez $lines_changed lignes modifiées. Pensez à effectuer un git push."
+            reccurency=$((reccurency + 1))
         fi
     else
         # Le code n'est pas assez important pour être push
